Dharma-Platform-Live-Dashboard-master/services/api-gateway-service/requirements.txt
# API Gateway Service Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Rate limiting
slowapi==0.1.9

# Database
asyncpg==0.29.0
redis==5.0.1

# HTTP client
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
structlog==23.2.0